Продолжена работа над аналоговым спидометром - нелинейная шкала; - цифровое отображение. Добавлено отображение уровня GPS.

+++ RaceData/Classes/Presentation Layer/UserStories/GPS/View/RDGPSViewInput.h
@@ -16,6 +16,6 @@
  Метод настраивает начальный стейт view
  */
 - (void)setupInitialState;
-- (void)updateSignalQualityLevel:(CGFloat)level;
+- (void)updateSignalQualityLevel:(NSUInteger)level;
 
 @end
